+++ src/Client/ClientCore/Systems/Input/InputSystem.cs
@@ -65,6 +65,7 @@ public class InputSystem : ISystem, IDisposable
             EventId.Client_Input_KeyUp,
             EventId.Client_Input_KeyDown,
             EventId.Client_Input_MouseMotion,
+            EventId.Client_Input_MouseWheel,
             EventId.Client_Network_PlayerEntitySelected,
             EventId.Core_Tick
         };
@@ -95,6 +96,7 @@ public class InputSystem : ISystem, IDisposable
 
                 // Route mouse events appropriately.
                 case EventId.Client_Input_MouseMotion:
+                case EventId.Client_Input_MouseWheel:
                     mouseEventHandler.HandleMouseEvent(ev);
                     break;